changed k from int to float, in harris detector

+++ include/detectors/harris.h
@@ -5,7 +5,7 @@
 
 void harrisKp(cv::Mat img, std::vector<KeyPoints> &kp, bool is_hdr = false,
               int msobel = SOBEL_SIZE, int mgauss = GAUSS_SIZE,
-              float sigma_x = SIGMA_X, float sigma_y = SIGMA_Y, int k = K,
+              float sigma_x = SIGMA_X, float sigma_y = SIGMA_Y, float k = K,
               float min_quality = MIN_QUALITY, int msize = MAXSUP_SIZE,
               int cv_size = CV_SIZE);
